linux-igep: rename recipe and update to latest version.
The name of linux-igep2 can produce confusion as seems it's only for the
IGEP v2 board, this patch renames the recipe to linux-igep that it's
a more generic name to include all the IGEP technology based boards.

The linux-igep kernel is provided by ISEE and supports all IGEP technology
based boards.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>