igep00x0: add u-boot config option to built with NAND support
IGEP-based boards can have two different flash memories and u-boot
uses different configuration options for each memory type.

If a user builds the first and second stage u-boot bootloader
binaries using the "igep0020_config" board configuration and
tries to use them on an IGEPv2 board with a NAND flash memory
then the kernel is not going to boot.

It's better to document on the igep00x0 machine configuration file
what the options are so users don't have to figure it out.

Signed-off-by: Javier Martinez Canillas <javier@dowhile0.org>