linux: igep00x0: built-in TFP410 support instead of as a module
Since the TFP410 DPI-to-DVI chip is used to display video on
IGEP boards and the igep00x0 defconfig is only used for these
devices, it is better to have it built in the kernel instead
than as a module.

Also, configure to have only one overlay instead of the maximum
three, since there is an bug in latest Xorg server that configures
the remaining overlay buffes with paddr equal to zero.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>