linux_git: igep00x0: Switch to use device tree based kernel.
As this package is only for development purposes, and we have initial
support for device tree in mainline, I think it's time to switch to
a device tree based model.

Note that this package is not enabled by default, and for that reason,
this update doesn't break anything. To use this recipe you should change
the DEFAULT_PREFERENCE to a bigger value.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>