igep00x0: Use specific machine names for every processor board.
It's right that all IGEP Processor Boards share a lot of stuff, but they can
differ as none all use the same processor. To make all of this more customizable
for every kind of processor board we should set a machine configuration file
for every IGEP Processor Board. To avoid duplication of code we also take care
of the common parts for all machines adding a shared igep00x0.inc file.

So, from now, these are the machines supported:

  * igep0020: for the IGEPv2 AM/DM37x Processor Board.
  * igep0030: for the IGEP COM MODULE AM/DM37x Processor Board.
  * igep0033: fot the IGEP COM AQUILA AM335x Processor Board.

Note that, with these patches, we also switch to use a hard floating point
flavour.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>